==> notebook/retrieval_evaluation/01_create_doc_ids_for_dataset.ipynb <==
## Create dataset with documents ids
The goal of this notebook is to merge all documents into single one and add the `document id` property to every document.

from pprint import pprint

import hashlib
import pandas as pd
import os

### Read all data files and merge them into a single one.

source_faqs_directory = "/home/jovyan/work/mage/data/faqs/"
doc_with_ids_path = "/home/jovyan/work/notebook/retrieval_evaluation/dataset_with_doc_ids.csv"

### Add a document id field
The length of the field is the same, as pipeline uses to index data.

fileset = []
column_names=["source_system", "category", "question", "answer"]

for file_name in os.listdir(source_faqs_directory):
    if file_name.endswith(".csv"):
        data = pd.read_csv(os.path.join(source_faqs_directory, file_name), delimiter=";", names=column_names, skiprows=1, usecols=(0,1,2,3))
        fileset.append(data)

df = pd.concat(fileset, ignore_index=True)
pprint(df[:1].to_dict(orient="records"))

for idx, row in df.iterrows():
    hex_digest = hashlib.md5(row["question"].encode() + row["answer"].encode()).hexdigest()
    id = hex_digest[:12]
    df.at[idx, "document_id"] = id

df[:1]

### Save the final dataframe

df.to_csv(doc_with_ids_path, sep=";", index=False)
