==> notebook/retrieval_evaluation/02_create_ground_truth.ipynb <==
## Create ground truth data
The goal of this notebook is to create a ground truth data to evaluate retrieval and RAG flow metrics.

from services import *
from common.settings import Settings
from common.client_factory import ClientFactory
import pandas as pd
from pandas import DataFrame
import json
import os
from typing import Any, Dict, List
from pprint import pprint
from tqdm import tqdm

### Read a dataset with documetns ids.

doc_with_ids_path = "/home/jovyan/work/notebook/retrieval_evaluation/dataset_with_doc_ids.csv"
ground_truth_path = "/home/jovyan/work/notebook/retrieval_evaluation/ground_truth.csv"

df = pd.read_csv(doc_with_ids_path, delimiter=";")
df[:1]

settings = Settings()
settings.to_dict()

factory = ClientFactory(settings)
es_client = factory.create_elasticsearch_client()
es_client.info()

### Create 5 questions to every question/answer pair
Create LLM client and request creating 5 questions for each question/answer pair.

bedrock_client = factory.create_bedrock_client()

def invoke_llm_model(prompt: str, bedrock_client)-> str:
    model = "amazon.titan-text-premier-v1:0"

    text_generation_config: dict[str, int | float | str] = {
        "maxTokenCount": 2000
    }

    body_request = json.dumps(
        {
            "inputText": prompt,
            "textGenerationConfig": text_generation_config
        }
    )

    response = bedrock_client.invoke_model(
        modelId=model,
        contentType="application/json",
        accept="*/*",
        body=body_request,
    )

    body_as_plain_text = response.get("body").read()
    response_body = json.loads(body_as_plain_text)
    return response_body
    # return response_body["results"][0]["outputText"].strip()

def create_prompt_for_document(document: Dict[str, Any]) -> str:
    question = document["question"].strip()
    answer = document["answer"].strip()
    category = document["category"].strip()
    
    prompt_en = (
        "Act like you are an experienced customer support service specialist at EV Digital Invest AG company. \n"
        "You are tasked to write five new questions for the given content.  \n"
        "Follow the instructions below to prepare the data. \n"

        "# Instructions \n"
        "- Read a question, answer, and the related category name from the internal database. \n"
        "- Analyze the question, answer, and the related category. \n"
        "- Based on analytical outcome, create 5 new questions. \n"
        "- Format this questions as CSV file with the semicolon as a delimiter in the format: \n"
        "category;question \n"
        "{category};{question number 1} \n"
        "{category};{question number 2} \n"
        "{category};{question number 3} \n"
        "{category};{question number 4} \n"
        "{category};{question number 5} \n"

        "# Question, answer, and the related category from the database: \n"
        "Question: {Question} \n"
        "Answer: {Answer} \n"
        "Category: {Category} \n"
    )

    prompt = (
        "Verhalte Dich, als wärst Du ein erfahrener Kundendienstspezialist bei der EV Digital Invest AG. \n"
        "Deine Aufgabe ist es, fünf neue Fragen zum gegebenen Inhalt zu formulieren. \n"
        "Befolge die untenstehenden Anweisungen, um die Daten vorzubereiten. \n"

        "# Anweisungen \n"
        "- Lies eine Frage, Antwort und den dazugehörigen Abschnittsnamen aus der internen Datenbank. \n"
        "- Analysiere die Frage, Antwort und den zugehörigen Abschnitt. \n"
        "- Erstelle basierend auf dem Analyseergebnis 5 neue Fragen. \n"
        "- Formatiere diese Fragen als CSV-Datei mit Semikolon als Trennzeichen im folgenden Format: \n"
        "Abschnitt;Frage \n"
        "{Abschnitt};{Frage Nummer 1} \n"
        "{Abschnitt};{Frage Nummer 2} \n"
        "{Abschnitt};{Frage Nummer 3} \n"
        "{Abschnitt};{Frage Nummer 4} \n"
        "{Abschnitt};{Frage Nummer 5} \n"

        "# Frage, Antwort und der zugehörige Abschnitt aus der Datenbank: \n"
        f"Frage: {question} \n"
        f"Antwort: {answer} \n"
        f"Abschnitt: {category} \n"
    )
    
    return prompt

#### Generate ground truth questions for the first item to test the approach

first_doc = df.iloc[0]
first_doc.to_dict()

prompt = create_prompt_for_document(first_doc.to_dict())
pprint(prompt)

generated_result = invoke_llm_model(prompt, bedrock_client)
generated_result

generated_questions = generated_result["results"][0]["outputText"].strip().split("\n")
for question in generated_questions:
    question_as_list = question.split(";")
    question_as_object = {
        "source_system": first_doc["source_system"],
        "category": question_as_list[0],
        "question": question_as_list[1],
        "document_id": first_doc["document_id"]
    }
    print(question_as_object)

#### Generate ground truth questions for all items

def generate_questions_for_prompt(doc: Dict[str, Any], prompt: str, bedrock_client) -> List[Dict[str, Any]]:
    results = []
    generated_result = invoke_llm_model(prompt, bedrock_client)
    generated_questions = generated_result["results"][0]["outputText"].strip().split("\n")

    for question in generated_questions:
        question_as_list = question.split(";")
        question_as_object = {
            "source_system": doc["source_system"],
            "category": question_as_list[0].strip(),
            "question": question_as_list[1].strip(),
            "document_id": doc["document_id"]
        }

        results.append(question_as_object)

    return results

existing_df: DataFrame
if os.path.exists(ground_truth_path):
    existing_df = pd.read_csv(ground_truth_path, delimiter=";")
else:
    columns = ['source_system', 'category', 'question', 'document_id']
    existing_df = pd.DataFrame(columns=columns)

existing_df[:10]

for idx, doc in tqdm(df.iterrows()):
    document_id_to_check = doc["document_id"]
    if not existing_df.empty and document_id_to_check in existing_df["document_id"].values:
        continue

    doc_as_dict = doc.to_dict()
    prompt = create_prompt_for_document(doc_as_dict)
    generated_questions = generate_questions_for_prompt(doc_as_dict, prompt, bedrock_client)

    documents_df = pd.DataFrame(generated_questions)
    existing_df = pd.concat([existing_df, documents_df], ignore_index=True)
    existing_df.to_csv(ground_truth_path, sep=";", index=False)

existing_df.head()

existing_df.tail()
